refactors

+++ src/main/java/com/github/bgalek/hackserver/application/player/web/RegistrationRequest.java
@@ -1,4 +1,4 @@
 package com.github.bgalek.hackserver.application.player.web;
 
-record RegistrationRequest(String name, String host, long port) {
+record RegistrationRequest(String name, String host, int port) {
 }
